refactor: move `ContextInitializer` into `context` module
